app: Rename source files to match purpose and read better in editor tabs
