+++ src/emulator/app/include/app/app_functions.h
