+++ src/emulator/app/src/app_config.cpp
@@ -15,7 +15,7 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-#include <app/config.h>
+#include <app/app_config.h>
 
 #include <host/version.h>
 #include <psp2/system_param.h>
