+++ src/emulator/app/src/app_private.h
