+++ src/emulator/app/src/app_sfo.cpp
@@ -15,7 +15,7 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-#include "sfo.h"
+#include "app_private.h"
 
 #include <host/sfo.h>
